requirements.txt
fastapi==0.115.13
mcp==1.9.4
pydantic==2.11.7
langchain==0.3.25
python-dotenv==1.1.0
psycopg2-binary==2.9.10
langchain-google-community==2.0.7
langchain-tavily==0.2.11
pandas==2.3.1
openai==1.98.0
PyJWT==2.10.1